Add user info scree, select photo of gallery and validate connectivity of internet

+++ lib/screens/brand_screen.dart
@@ -1,4 +1,5 @@
 import 'package:adaptive_dialog/adaptive_dialog.dart';
+import 'package:connectivity/connectivity.dart';
 import 'package:flutter/material.dart';
 
 import 'package:vehicles_app/components/loader_component.dart';
@@ -166,6 +167,24 @@ class _BrandScreenState extends State<BrandScreen> {
       _showLoader = true;
     });
 
+    // validar la conexión a internet
+    var connectivityResult = await Connectivity().checkConnectivity();
+    if (connectivityResult == ConnectivityResult.none) {
+      setState(() {
+        _showLoader = false;
+      });
+
+      await showAlertDialog(
+          context: context,
+          title: 'Error',
+          message: 'Verifica que estés conectado a internet.',
+          actions: <AlertDialogAction>[
+            AlertDialogAction(key: null, label: 'Aceptar')
+          ]);
+
+      return;
+    }
+
     Map<String, dynamic> request = {
       'id': widget.brand.id,
       'description': _description
@@ -211,6 +230,24 @@ class _BrandScreenState extends State<BrandScreen> {
     setState(() {
       _showLoader = true;
     });
+
+    // validar la conexión a internet
+    var connectivityResult = await Connectivity().checkConnectivity();
+    if (connectivityResult == ConnectivityResult.none) {
+      setState(() {
+        _showLoader = false;
+      });
+
+      await showAlertDialog(
+          context: context,
+          title: 'Error',
+          message: 'Verifica que estés conectado a internet.',
+          actions: <AlertDialogAction>[
+            AlertDialogAction(key: null, label: 'Aceptar')
+          ]);
+
+      return;
+    }
 
     Response response = await ApiHelper.delete(
         '/api/Brands/', widget.brand.id.toString(), widget.token.token);
