+++ lib/screens/user_screen.dart
@@ -1,9 +1,12 @@
+import 'dart:convert';
 import 'dart:io';
 
 import 'package:adaptive_dialog/adaptive_dialog.dart';
 import 'package:camera/camera.dart';
+import 'package:connectivity/connectivity.dart';
 import 'package:email_validator/email_validator.dart';
 import 'package:flutter/material.dart';
+import 'package:image_picker/image_picker.dart';
 
 import 'package:vehicles_app/components/loader_component.dart';
 import 'package:vehicles_app/helpers/api_helper.dart';
@@ -256,15 +259,40 @@ class _UserScreenState extends State<UserScreen> {
       _showLoader = true;
     });
 
+    // validar la conexión a internet
+    var connectivityResult = await Connectivity().checkConnectivity();
+    if (connectivityResult == ConnectivityResult.none) {
+      setState(() {
+        _showLoader = false;
+      });
+
+      await showAlertDialog(
+          context: context,
+          title: 'Error',
+          message: 'Verifica que estés conectado a internet.',
+          actions: <AlertDialogAction>[
+            AlertDialogAction(key: null, label: 'Aceptar')
+          ]);
+
+      return;
+    }
+
+    String base64image = '';
+    if (_photoChanged) {
+      List<int> imageBytes = await _image.readAsBytes();
+      base64image = base64Encode(imageBytes);
+    }
+
     Map<String, dynamic> request = {
       'firstName': _firstName,
       'lastName': _lastName,
-      'documentType': _documentTypeId,
+      'documentTypeId': _documentTypeId,
       'document': _document,
       'email': _email,
       'userName': _email,
       'address': _address,
-      'phoneNumber': _phoneNumber
+      'phoneNumber': _phoneNumber,
+      'image': base64image,
     };
 
     Response response =
@@ -293,16 +321,41 @@ class _UserScreenState extends State<UserScreen> {
       _showLoader = true;
     });
 
+    // validar la conexión a internet
+    var connectivityResult = await Connectivity().checkConnectivity();
+    if (connectivityResult == ConnectivityResult.none) {
+      setState(() {
+        _showLoader = false;
+      });
+
+      await showAlertDialog(
+          context: context,
+          title: 'Error',
+          message: 'Verifica que estés conectado a internet.',
+          actions: <AlertDialogAction>[
+            AlertDialogAction(key: null, label: 'Aceptar')
+          ]);
+
+      return;
+    }
+
+    String base64image = '';
+    if (_photoChanged) {
+      List<int> imageBytes = await _image.readAsBytes();
+      base64image = base64Encode(imageBytes);
+    }
+
     Map<String, dynamic> request = {
       'id': widget.user.id,
       'firstName': _firstName,
       'lastName': _lastName,
-      'documentType': _documentTypeId,
+      'documentTypeId': _documentTypeId,
       'document': _document,
       'email': _email,
       'userName': _email,
       'address': _address,
-      'phoneNumber': _phoneNumber
+      'phoneNumber': _phoneNumber,
+      'image': base64image,
     };
 
     Response response = await ApiHelper.put(
@@ -346,6 +399,24 @@ class _UserScreenState extends State<UserScreen> {
       _showLoader = true;
     });
 
+    // validar la conexión a internet
+    var connectivityResult = await Connectivity().checkConnectivity();
+    if (connectivityResult == ConnectivityResult.none) {
+      setState(() {
+        _showLoader = false;
+      });
+
+      await showAlertDialog(
+          context: context,
+          title: 'Error',
+          message: 'Verifica que estés conectado a internet.',
+          actions: <AlertDialogAction>[
+            AlertDialogAction(key: null, label: 'Aceptar')
+          ]);
+
+      return;
+    }
+
     Response response = await ApiHelper.delete(
         '/api/Users/', widget.user.id, widget.token.token);
 
@@ -368,34 +439,34 @@ class _UserScreenState extends State<UserScreen> {
   }
 
   Widget _showPhoto() {
-    return InkWell(
-      onTap: () => _takePicture(),
-      child: Stack(children: <Widget>[
-        Container(
-          margin: EdgeInsets.only(top: 10),
-          child: widget.user.id.isEmpty && !_photoChanged
-              ? Image(
-                  image: AssetImage('assets/noimage.png'),
-                  height: 160,
-                  width: 160,
-                  fit: BoxFit.cover)
-              : ClipRRect(
-                  borderRadius: BorderRadius.circular(80),
-                  child: _photoChanged
-                      ? Image.file(File(_image.path),
-                          width: 160, height: 160, fit: BoxFit.cover)
-                      : FadeInImage(
-                          placeholder: AssetImage('assets/vehicles_logo.jpg'),
-                          image: NetworkImage(widget.user.imageFullPath),
-                          width: 160,
-                          height: 160,
-                          fit: BoxFit.cover,
-                        ),
-                ),
-        ),
-        Positioned(
-            bottom: 0,
-            left: 100,
+    return Stack(children: <Widget>[
+      Container(
+        margin: EdgeInsets.only(top: 10),
+        child: widget.user.id.isEmpty && !_photoChanged
+            ? Image(
+                image: AssetImage('assets/noimage.png'),
+                height: 160,
+                width: 160,
+                fit: BoxFit.cover)
+            : ClipRRect(
+                borderRadius: BorderRadius.circular(80),
+                child: _photoChanged
+                    ? Image.file(File(_image.path),
+                        width: 160, height: 160, fit: BoxFit.cover)
+                    : FadeInImage(
+                        placeholder: AssetImage('assets/vehicles_logo.jpg'),
+                        image: NetworkImage(widget.user.imageFullPath),
+                        width: 160,
+                        height: 160,
+                        fit: BoxFit.cover,
+                      ),
+              ),
+      ),
+      Positioned(
+          bottom: 0,
+          left: 100,
+          child: InkWell(
+            onTap: () => _takePicture(),
             child: ClipRRect(
                 borderRadius: BorderRadius.circular(30),
                 child: Container(
@@ -403,9 +474,22 @@ class _UserScreenState extends State<UserScreen> {
                     height: 60,
                     width: 60,
                     child: Icon(Icons.photo_camera,
-                        size: 40, color: Colors.blue))))
-      ]),
-    );
+                        size: 40, color: Colors.blue))),
+          )),
+      Positioned(
+          bottom: 0,
+          left: 0,
+          child: InkWell(
+            onTap: () => _selectPicture(),
+            child: ClipRRect(
+                borderRadius: BorderRadius.circular(30),
+                child: Container(
+                    color: Colors.green[50],
+                    height: 60,
+                    width: 60,
+                    child: Icon(Icons.image, size: 40, color: Colors.blue))),
+          ))
+    ]);
   }
 
   Widget _showLastName() {
@@ -473,6 +557,7 @@ class _UserScreenState extends State<UserScreen> {
     return Container(
       padding: EdgeInsets.all(10),
       child: TextField(
+        enabled: widget.user.id.isEmpty,
         keyboardType: TextInputType.emailAddress,
         controller: _emailController,
         decoration: InputDecoration(
@@ -534,6 +619,24 @@ class _UserScreenState extends State<UserScreen> {
       _showLoader = true;
     });
 
+    // validar la conexión a internet
+    var connectivityResult = await Connectivity().checkConnectivity();
+    if (connectivityResult == ConnectivityResult.none) {
+      setState(() {
+        _showLoader = false;
+      });
+
+      await showAlertDialog(
+          context: context,
+          title: 'Error',
+          message: 'Verifica que estés conectado a internet.',
+          actions: <AlertDialogAction>[
+            AlertDialogAction(key: null, label: 'Aceptar')
+          ]);
+
+      return;
+    }
+
     Response response = await ApiHelper.getDocumentTypes(widget.token.token);
 
     setState(() {
@@ -582,6 +685,17 @@ class _UserScreenState extends State<UserScreen> {
       setState(() {
         _photoChanged = true;
         _image = response.result;
+      });
+    }
+  }
+
+  void _selectPicture() async {
+    final ImagePicker _picker = ImagePicker();
+    final XFile? image = await _picker.pickImage(source: ImageSource.gallery);
+    if (image != null) {
+      setState(() {
+        _photoChanged = true;
+        _image = image;
       });
     }
   }
